% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagsTextInput.R
\name{tagsTextInput}
\alias{tagsTextInput}
\title{text input specific to tags}
\usage{
tagsTextInput(...)
}
\arguments{
\item{...}{see \code{\link[shiny]{textInput}}}
}
\description{
text input specific to tags
}
\examples{
\dontrun{
  library(shiny)
  ui <- fluidPage(
    tagsTextInput("fruits", "Fruits", "apple, banana"),
    textOutput("out")
  )

  server <- function(input, output){
    output$out <- renderPrint( strsplit( input$fruits, ",")[[1]] )
  }

  shinyApp( ui, server )

}
}
